Force students to select sane settings
Check for how many chords are possible with the given settings.
Under a certain number of chords, output a warning alert dialog.
With zero chords, output an error alert that doesn't go away.
Check when the app is first launched and whenever it is resumed.
Fix bugs with weird range edge cases, like backwards ranges.